requirements.txt
streamlit
streamlit-extras
streamlit-lottie
streamlit-option-menu
plotly-express
pymongo
pandas
seaborn